MOON-285: Styling updates
