+++ src/components/Table/TableRow/TableRow.scss
@@ -4,7 +4,7 @@ $row-multiple-line-height: 64px;
 .moonstone-tableRow {
     height: $row-regular-height;
 
-    border-bottom: 1px solid var(--color-gray40);
+    border-bottom: 1px solid var(--color-gray_light);
 
     transition: background-color 0.2s ease-in-out;
 
@@ -12,13 +12,31 @@ $row-multiple-line-height: 64px;
         background-color: var(--color-gray_light40);
     }
 
-    &.moonstone-tableRow-selected {
+    &.moonstone-tableRow-highlighted {
         color: var(--color-light);
 
         background-color: var(--color-accent);
 
         &:hover {
             background-color: var(--color-accent);
+        }
+
+        .moonstone-checkbox_input {
+            border: 1px solid var(--color-light);
+
+            &[aria-checked="false"] {
+                box-shadow: none;
+            }
+
+            &.hasFocus {
+                border: 2px solid var(--color-accent_dark_contrast);
+            }
+        }
+
+        .moonstone-chip {
+            color: var(--color-light);
+
+            background-color: var(--color-accent_light40);
         }
     }
