+++ src/components/Table/TableRow/TableRow.types.ts
@@ -9,12 +9,19 @@ export type TableRowProps = {
     /**
      * Whether the cell height should be increased to show more than 1 line
      */
-    isMultipleLines?: boolean;
+    hasMultipleLines?: boolean;
 
     /**
      * If true, then the row is selected
      */
     isSelected?: boolean;
+
+    /**
+     * If true, then the row is highlighted. The background color of the row
+     * will be in contrast to the normal row color and supported components
+     * nested within the row will have their style change correspondingly
+     */
+    isHighlighted?: boolean;
 
     /**
      * The children elements
