--- src/data/tableDataFlat.ts
@@ -1,38 +1,0 @@
-export const tableDataFlat = [
-    {
-        name: 'Demo Roles and Users',
-        type: 'Page',
-        createdBy: 'root',
-        lastModifiedOn: 'Aug. 12, 2016',
-    },
-    {
-        name: 'About',
-        type: 'Page',
-        createdBy: 'root',
-        lastModifiedOn: 'Jan. 6, 2016',
-    },
-    {
-        name: 'Search Results',
-        type: 'Page',
-        createdBy: 'system',
-        lastModifiedOn: 'Feb. 29, 2016',
-    },
-    {
-        name: 'Corporate Responsibility',
-        type: 'Page',
-        createdBy: 'root',
-        lastModifiedOn: 'Jan. 26, 2016',
-    },
-    {
-        name: 'How to Use This Demo',
-        type: 'Page',
-        createdBy: 'root',
-        lastModifiedOn: 'Jan. 26, 2016',
-    },
-    {
-        name: 'Our Companies',
-        type: 'Page',
-        createdBy: 'root',
-        lastModifiedOn: 'Jan. 26, 2016',
-    }
-];
